refactor tooltips for 3d
