+++ app/packages/state/src/hooks/index.ts
@@ -16,6 +16,7 @@ export { default as useSetExtendedSelection } from "./useSetExtendedSelection";
 export { default as useSetSelectedLabels } from "./useSetSelectedLabels";
 export { default as useSetView } from "./useSetView";
 export { default as useStateUpdate } from "./useStateUpdate";
+export { default as useTooltip } from "./useTooltip";
 export * from "./useOnSelectLabel";
 export * from "./useStateUpdate";
 export { default as useTo } from "./useTo";
